pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "a bot for scheduling messages"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=20.0",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.0",
]
